# docs/swagger.yaml
basePath: /
definitions:
...
  api.ChangePasswordReq:
    properties:
      newPassword:
        minLength: 6
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  api.CreateBotReq:
    properties:
      id:
        type: string
      llmLight:
        type: string
      llmMain:
        type: string
      maxTokens:
        type: integer
      model:
        type: string
      name:
        type: string
      systemPrompt:
        type: string
      temperature:
        type: number
      workers:
        type: integer
    required:
    - id
    - name
    type: object
  api.CreateChannelReq:
    properties:
      config:
        description: JSON 字符串
        type: string
      name:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
...
  api.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.LoginResp:
    properties:
      avatar:
        type: string
      displayName:
        type: string
      id:
        type: integer
      lastLoginAt:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  api.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  api.UpdateBotReq:
    properties:
      llmLight:
        type: string
      llmMain:
        type: string
      maxTokens:
        type: integer
      model:
        type: string
      name:
        type: string
      systemPrompt:
        type: string
      temperature:
        type: number
      workers:
        type: integer
    type: object
  api.UpdateChannelReq:
    properties:
      config:
        type: string
      enabled:
        type: boolean
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 多渠道 AI 聊天机器人框架的 HTTP API
  title: thinkbot API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 通过用户名密码登录，返回 JWT Cookie
      parameters:
      - description: 登录请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.LoginResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Response'
      summary: 用户登录
      tags:
      - 认证
...
  /api/auth/me:
    get:
      description: 返回当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.LoginResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 获取当前用户
      tags:
      - 认证
  /api/auth/password:
    put:
      consumes:
      - application/json
      description: 当前用户修改自己的登录密码
      parameters:
      - description: 修改密码请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 修改密码
      tags:
      - 认证
...
  /api/bots:
    get:
      description: 列出所有 Bot 定义及其运行状态
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: Bot 列表
      tags:
      - Bot 管理
    post:
      consumes:
      - application/json
      description: 创建新的 Bot 定义（需要 bot.manage 权限）
      parameters:
      - description: 创建 Bot 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateBotReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 创建 Bot
      tags:
      - Bot 管理
  /api/bots/{id}:
    delete:
      description: 删除指定 Bot 定义
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 删除 Bot
      tags:
      - Bot 管理
    get:
      description: 获取指定 Bot 的定义和运行时信息
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 获取 Bot
      tags:
      - Bot 管理
    put:
      consumes:
      - application/json
      description: 更新指定 Bot 的配置（字段可选，需要 bot.manage 权限）
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新 Bot 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateBotReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 更新 Bot
      tags:
      - Bot 管理
  /api/bots/{id}/channels:
    get:
      description: 返回指定 Bot 的所有 Channel 配置
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: Channel 列表
      tags:
      - Channel 管理
    post:
      consumes:
      - application/json
      description: 为指定 Bot 创建新的 Channel 配置
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 创建 Channel 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateChannelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 创建 Channel
      tags:
      - Channel 管理
  /api/bots/{id}/channels/{cid}:
    delete:
      description: 删除指定的 Channel 配置
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel ID
        in: path
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 删除 Channel
      tags:
      - Channel 管理
    put:
      consumes:
      - application/json
      description: 更新指定的 Channel 配置
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel ID
        in: path
        name: cid
        required: true
        type: string
      - description: 更新 Channel 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateChannelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 更新 Channel
      tags:
      - Channel 管理
...
  /api/bots/{id}/start:
    post:
      description: 启动指定 Bot
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 启动 Bot
      tags:
      - Bot 管理
  /api/bots/{id}/stop:
    post:
      description: 停止指定 Bot
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 停止 Bot
      tags:
      - Bot 管理
  /api/channels/types:
    get:
      description: 返回系统支持的 Channel 类型及其配置字段
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: Channel 类型
      tags:
      - Channel 管理
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"